Create DrawCircle.cpp
Description
There are N points in a plane, each with a two-dimensional coordinate. We want to draw a circle which has at least two points within or on its border. You are asked to calculate the minimum radius of this circle.

Input
The first line is an integer N, denoting the number of points. Following are N lines. Each contains a pair of real number (x, y), denoting the coordinate of a point.|x|<=10^9,|y|<=10^9.
40% of the test cases guarantee that N <= 1000.
All test cases guarantee that N <= 200000.

Output
Output only a real number rounding to 3 decimals after decimal point, denoting the minimum radius.