Create DNASequenceAlignment.c
Description
Bob is attracted in Bioinformatics recently. He reads papers day and night and devotes all his mind in studying. Today he is going to review the basic problem in Bioinformatics: DNA sequence alignment. His purpose is to find a simple and effective algorithm that performs global alignment with two highly similar DNA sequences.

A DNA sequence is presented as a sequence of characters, which may be 'A', 'G', 'C' or 'T'. To align two DNA sequences, some gaps may be inserted to sequences so that two sequences have the same length. And then it is counted up for every pair of matched characters by a score matrix. Bob uses a minimal-score matrix hence the total score of alignment should be minimized. Following is the score matrix he uses:

        A	T	C	G	-
A	0	5	5	5	3
T	5	0	5	5	3
C	5	5	0	4	3
G	5	5	4	0	3
-	3	3	3	3	3
​For example, an alignment for DNA sequences "AAGACG" and "CAGAGCTC" may be:
-AAGA-C-G
CA-GAGCTC
The total score is 3+0+3+0+0+3+0+3+4=16.

Input
Input contains two lines, which are the two DNA sequences to align. DNA sequences contain only characters 'A', 'G', 'C' and 'T'. The length of each sequence is not greater than 1000.

Output
Output contains one integer which is the minimal total score of alignment.

Sample Input 1 
AGTGCTGAAAGTTGCGCCAGTGAC
AGTGCTGAAGTTCGCCAGTTGACG

Sample Output 1
12